+++ tests/test_main.py
@@ -74,10 +74,19 @@ class TestActivateUninstall(unittest.TestCase):
         self.mock_product_catalog_cls = patch('sat_install_utility.main.ProductCatalog').start()
         self.mock_product_catalog = self.mock_product_catalog_cls.return_value
         self.fake_products = [
-            Mock(product_name='sat', version='2.0.3', docker_image_name='cray/cray-sat', docker_image_version='3.4.0'),
-            Mock(product_name='sat', version='2.0.4', docker_image_name='cray/cray-sat', docker_image_version='3.5.0'),
-            Mock(product_name='cos', version='2.0.5', docker_image_name='cray/cray-cos', docker_image_version='3.4.0'),
-            Mock(product_name='cos', version='2.0.6', docker_image_name='cray/cray-cos', docker_image_version='3.5.0')
+            # Multiple versions of SAT that have some images in common with one another.
+            Mock(product_name='sat', version='2.0.3', docker_images={'cray/cray-sat': '3.4.0',
+                                                                     'cray/sat-cfs-install': '1.2.3'}),
+            Mock(product_name='sat', version='2.0.4', docker_images={'cray/cray-sat': '3.5.0'}),
+            Mock(product_name='sat', version='2.2.5', docker_images={'cray/cray-sat': '3.5.0',
+                                                                     'cray/sat-cfs-install': '1.2.3',
+                                                                     'cray/sat-install-utility': '1.1.0'}),
+            # Two versions of COS with different images, but same image version numbers as the SAT products above.
+            Mock(product_name='cos', version='2.0.5', docker_images={'cray/cray-cos': '3.4.0'}),
+            Mock(product_name='cos', version='2.0.6', docker_images={'cray/cray-cos': '3.5.0'}),
+            # Two products that each have a single docker image, and are the same image.
+            Mock(product_name='sat', version='2.0.1', docker_images={'cray/cray-sat': '3.1.0'}),
+            Mock(product_name='another-product', version='1.2.3', docker_images={'cray/cray-sat': '3.1.0'})
         ]
         [p.configure_mock(name=p.product_name) for p in self.fake_products]
 
@@ -108,33 +117,55 @@ class TestActivateUninstall(unittest.TestCase):
         )
         product = self.fake_products[0]
         product.uninstall_hosted_repo.assert_called_once_with(self.mock_nexus_api, 'sle-15sp3')
-        product.uninstall_docker_image.assert_called_once_with(self.mock_docker_api)
+        product.uninstall_docker_image.assert_called_once_with('cray/cray-sat', '3.4.0', self.mock_docker_api)
+        product.remove_from_product_catalog.assert_called_once_with('cray-product-catalog', 'services')
 
     def test_uninstall_without_docker_image_removal(self):
         """Test a successful uninstall when a Docker image is not removed."""
-        shared_image_product = Mock(
-            product_name='some-other-product',
-            version='1.2.3',
-            docker_image_name='cray/cray-sat',
-            docker_image_version='3.4.0'
-        )
-        shared_image_product.configure_mock(name=shared_image_product.product_name)
-        self.fake_products.append(shared_image_product)
         uninstall(
             name='sat',
-            version='2.0.3',
+            version='2.0.1',
             dist='sle-15sp3',
             nexus_api=self.mock_nexus_api,
             docker_api=self.mock_docker_api,
         )
-        product = self.fake_products[0]
+        product = self.fake_products[-2]
         product.uninstall_hosted_repo.assert_called_once_with(self.mock_nexus_api, 'sle-15sp3')
         product.uninstall_docker_image.assert_not_called()
         expected_print_output = (
-            f'Not removing Docker image cray/cray-sat:3.4.0 '
-            f'used by the following other product versions: {str(shared_image_product)}'
+            f'Not removing Docker image cray/cray-sat:3.1.0 '
+            f'used by the following other product versions: {str(self.fake_products[-1])}'
         )
         self.mock_print.assert_called_with(expected_print_output)
+        product.remove_from_product_catalog.assert_called_once_with('cray-product-catalog', 'services')
+
+    def test_uninstall_with_partial_docker_image_removal(self):
+        """Test uninstalling a version where one of the Docker images is shared."""
+        uninstall(
+            name='sat',
+            version='2.2.5',
+            dist='sle-15sp3',
+            nexus_api=self.mock_nexus_api,
+            docker_api=self.mock_docker_api,
+        )
+        product = self.fake_products[2]
+        product.uninstall_hosted_repo.assert_called_once_with(self.mock_nexus_api, 'sle-15sp3')
+        expected_print_output = [
+            (
+                f'Not removing Docker image cray/cray-sat:3.5.0 '
+                f'used by the following other product versions: {str(self.fake_products[1])}'
+            ),
+            (
+                f'Not removing Docker image cray/sat-cfs-install:1.2.3 '
+                f'used by the following other product versions: {str(self.fake_products[0])}'
+            )
+        ]
+        for output in expected_print_output:
+            self.mock_print.assert_any_call(output)
+        product.uninstall_docker_image.assert_called_once_with(
+            'cray/sat-install-utility', '1.1.0', self.mock_docker_api
+        )
+        product.remove_from_product_catalog.assert_called_once_with('cray-product-catalog', 'services')
 
     def test_activate_success(self):
         """Test a successful activation."""