+++ tests/test_products.py
@@ -22,8 +22,9 @@ ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
 OTHER DEALINGS IN THE SOFTWARE.
 """
 
+from subprocess import CalledProcessError
 import unittest
-from unittest.mock import Mock
+from unittest.mock import Mock, patch
 
 from yaml import safe_dump
 
@@ -105,34 +106,80 @@ class TestInstalledProductVersion(unittest.TestCase):
     def setUp(self):
         """Set up mocks."""
         self.installed_product_version = InstalledProductVersion(
-            'sat', '2.0.0', {'component_versions': {'sat': '1.0.0'}}
+            'sat',
+            '2.2.0',
+            {
+                'component_versions': {
+                    'docker': {
+                        'cray/cray-sat': '3.9.0',
+                        'cray/sat-cfs-install': '1.2.0',
+                        'cray/sat-install-utility': '3.9.0',
+                    },
+                    'rpm': {
+                        'cray-sat-podman': '1.5.0'
+                    }
+                }
+            }
         )
+        self.legacy_installed_product_version = InstalledProductVersion(
+            'sat', '1.0.1', {'component_versions': {'sat': '1.0.0'}}
+        )
+
         self.mock_nexus_api = Mock()
-        self.mock_group_members = ['sat-3.0.0-sle-15sp3', 'sat-2.0.0-sle-15sp3', 'sat-1.0.1-sle-15sp3']
+        self.mock_group_members = ['sat-3.0.0-sle-15sp3', 'sat-2.2.0-sle-15sp3', 'sat-1.0.1-sle-15sp3']
         self.mock_group_repo = Mock()
         self.mock_group_repo.group.member_names = self.mock_group_members
         # This is slightly incorrect as NexusApi.repos.list may also return a
         # hosted repo, but good enough for the test.
         self.mock_nexus_api.repos.list.return_value = [self.mock_group_repo]
         self.mock_docker_api = Mock()
+        self.mock_environ = patch('sat_install_utility.products.os.environ').start()
+        self.mock_check_output = patch('sat_install_utility.products.subprocess.check_output').start()
 
-    def test_docker_image_version(self):
-        """Test getting the Docker image version."""
-        expected_docker_image_version = '1.0.0'
+    def tearDown(self):
+        """Stop patches."""
+        patch.stopall()
+
+    def test_docker_images(self):
+        """Test getting the Docker images."""
+        expected_docker_image_versions = {'cray/cray-sat': '3.9.0',
+                                          'cray/sat-cfs-install': '1.2.0',
+                                          'cray/sat-install-utility': '3.9.0'}
         self.assertEqual(
-            expected_docker_image_version, self.installed_product_version.docker_image_version
+            expected_docker_image_versions, self.installed_product_version.docker_images
         )
 
-    def test_docker_image_name(self):
-        """Test getting the Docker image name."""
-        expected_docker_image_name = 'cray/cray-sat'
+    def test_legacy_docker_images(self):
+        """Test getting the Docker images from an 'old'-style product catalog entry."""
+        expected_docker_image_versions = {'cray/cray-sat': '1.0.0'}
         self.assertEqual(
-            expected_docker_image_name, self.installed_product_version.docker_image_name
+            expected_docker_image_versions, self.legacy_installed_product_version.docker_images
         )
+
+    def test_no_docker_images(self):
+        """Test a product that has an empty dictionary under the 'docker' key returns an empty dictionary."""
+        product_with_no_docker_images = InstalledProductVersion(
+            'sat', '0.9.9', {'component_versions': {'docker': {}}}
+        )
+        self.assertEqual(product_with_no_docker_images.docker_images, {})
+
+    def test_no_docker_images_null(self):
+        """Test a product that has None under the 'docker' key returns an empty dictionary."""
+        product_with_no_docker_images = InstalledProductVersion(
+            'sat', '0.9.9', {'component_versions': {'docker': None}}
+        )
+        self.assertEqual(product_with_no_docker_images.docker_images, {})
+
+    def test_no_docker_images_empty_list(self):
+        """Test a product that has an empty list under the 'docker' key returns an empty dictionary."""
+        product_with_no_docker_images = InstalledProductVersion(
+            'sat', '0.9.9', {'component_versions': {'docker': []}}
+        )
+        self.assertEqual(product_with_no_docker_images.docker_images, {})
 
     def test_str(self):
         """Test the string representation of InstalledProductVersion."""
-        expected_str = 'sat-2.0.0'
+        expected_str = 'sat-2.2.0'
         self.assertEqual(
             expected_str, str(self.installed_product_version)
         )
@@ -146,7 +193,7 @@ class TestInstalledProductVersion(unittest.TestCase):
 
     def test_get_hosted_repo_name(self):
         """Test getting a hosted repo name for an InstalledProductVersion."""
-        expected_hosted_repo_name = 'sat-2.0.0-sle-15sp3'
+        expected_hosted_repo_name = 'sat-2.2.0-sle-15sp3'
         self.assertEqual(
             expected_hosted_repo_name, self.installed_product_version.get_hosted_repo_name('sle-15sp3')
         )
@@ -154,22 +201,44 @@ class TestInstalledProductVersion(unittest.TestCase):
     def test_uninstall_hosted_repo(self):
         """Test uninstalling a hosted repo for an InstalledProductVersion."""
         self.installed_product_version.uninstall_hosted_repo(self.mock_nexus_api, 'sle-15sp3')
-        self.mock_nexus_api.repos.delete.assert_called_once_with('sat-2.0.0-sle-15sp3')
+        self.mock_nexus_api.repos.delete.assert_called_once_with('sat-2.2.0-sle-15sp3')
 
     def test_uninstall_docker_image(self):
         """Test uninstalling a Docker image for an InstalledProductVersion."""
-        self.installed_product_version.uninstall_docker_image(self.mock_docker_api)
-        self.mock_docker_api.delete_image.assert_called_once_with('cray/cray-sat', '1.0.0')
+        self.installed_product_version.uninstall_docker_image('foo', 'bar', self.mock_docker_api)
+        self.mock_docker_api.delete_image.assert_called_once_with('foo', 'bar')
 
     def test_activate_hosted_repo(self):
+        """Test activating a product version's hosted repository."""
         self.installed_product_version.activate_hosted_repo(self.mock_nexus_api, 'sle-15sp3')
         self.mock_nexus_api.repos.raw_group.update.assert_called_once_with(
             self.mock_group_repo.name,
             self.mock_group_repo.online,
             self.mock_group_repo.storage.blobstore_name,
             self.mock_group_repo.storage.strict_content_type_validation,
-            member_names=('sat-2.0.0-sle-15sp3', 'sat-3.0.0-sle-15sp3', 'sat-1.0.1-sle-15sp3')
+            member_names=('sat-2.2.0-sle-15sp3',)
         )
+
+    def test_remove_from_product_catalog(self):
+        """Test removing a version from the product catalog."""
+        self.installed_product_version.remove_from_product_catalog('mock-name', 'mock-namespace')
+        self.mock_environ.update.assert_called_once_with({
+            'PRODUCT': self.installed_product_version.name,
+            'PRODUCT_VERSION': self.installed_product_version.version,
+            'CONFIG_MAP': 'mock-name',
+            'CONFIG_MAP_NS': 'mock-namespace'
+        })
+        self.mock_check_output.assert_called_once_with(['catalog_delete.py'])
+
+    def test_remove_from_product_catalog_fail(self):
+        """Test removing a version from the product catalog when the subcommand fails."""
+        expected_err_regex = (
+            f'Error removing {self.installed_product_version.name}-'
+            f'{self.installed_product_version.version} from product catalog'
+        )
+        self.mock_check_output.side_effect = CalledProcessError(1, 'catalog_delete.py')
+        with self.assertRaisesRegex(ProductInstallException, expected_err_regex):
+            self.installed_product_version.remove_from_product_catalog('mock-name', 'mock-namespace')
 
 
 if __name__ == '__main__':